Move type conversion to AbstractCommon and convert value keys.
