+++ pydov/types/abstract.py
@@ -1,16 +1,15 @@
 # -*- coding: utf-8 -*-
 """Module containing the base DOV data types."""
 
-import datetime
 import types
 import warnings
 from collections import OrderedDict
-from distutils.util import strtobool
 
 import pydov
 import numpy as np
 
 from owslib.etree import etree
+from pydov.search.abstract import AbstractCommon
 from pydov.util.dovutil import (
     get_dov_xml,
     parse_dov_xml,
@@ -23,7 +22,7 @@ from ..util.errors import (
 )
 
 
-class AbstractCommon(object):
+class AbstractTypeCommon(AbstractCommon):
     """Class grouping methods common to AbstractDovType and
     AbstractDovSubType."""
 
@@ -52,34 +51,6 @@ class AbstractCommon(object):
             `xpath`, converted to the type described by `returntype`.
 
         """
-        if returntype == 'string':
-            def typeconvert(x):
-                return x.strip()
-        elif returntype == 'integer':
-            def typeconvert(x):
-                return int(x)
-        elif returntype == 'float':
-            def typeconvert(x):
-                return float(x)
-        elif returntype == 'date':
-            def typeconvert(x):
-                # Patch for Zulu-time issue of geoserver for WFS 1.1.0
-                if x.endswith('Z'):
-                    return datetime.datetime.strptime(x, '%Y-%m-%dZ').date() \
-                           + datetime.timedelta(days=1)
-                else:
-                    return datetime.datetime.strptime(x, '%Y-%m-%d').date()
-        elif returntype == 'datetime':
-            def typeconvert(x):
-                return datetime.datetime.strptime(
-                    x.split('.')[0], '%Y-%m-%dT%H:%M:%S')
-        elif returntype == 'boolean':
-            def typeconvert(x):
-                return strtobool(x) == 1
-        else:
-            def typeconvert(x):
-                return x
-
         if namespace is not None:
             ns = '{%s}' % namespace
             text = func('./' + ns + ('/' + ns).join(xpath.split('/')))
@@ -88,10 +59,10 @@ class AbstractCommon(object):
 
         if text is None:
             return np.nan
-        return typeconvert(text)
 
+        return cls._typeconvert(text, returntype)
 
-class AbstractDovSubType(AbstractCommon):
+class AbstractDovSubType(AbstractTypeCommon):
 
     _name = None
     _rootpath = None
@@ -244,7 +215,7 @@ class AbstractDovSubType(AbstractCommon):
         return cls._rootpath
 
 
-class AbstractDovType(AbstractCommon):
+class AbstractDovType(AbstractTypeCommon):
     """Abstract DOV type grouping fields and methods common to all DOV
     object types. Not to be instantiated or used directly."""
